shellcheck fix

+++ ci/cloudbuild/builds/cmake-install.sh
@@ -150,7 +150,7 @@ quickstart::run_cmake_and_make "${INSTALL_PREFIX}"
 
 # Deletes all the installed artifacts, and installs only the runtime components
 # to verify that we can still execute the compiled quickstart programs.
-rm -rf "${INSTALL_PREFIX}"/{include,lib64}
+rm -rf "${INSTALL_PREFIX:?}"/{include,lib64}
 cmake --install cmake-out --component google_cloud_cpp_runtime
 quickstart::run_cmake_and_make "${INSTALL_PREFIX}"
